initial commit: split the base policy and the architectures
