+++ gflownet/policy/base.py
@@ -46,59 +46,6 @@ class ModelBase(ABC):
     def __call__(self, states):
         return self.model(states)
 
-    def make_mlp(self, activation):
-        """
-        Defines an MLP with no top layer activation
-        If share_weight == True,
-            baseModel (the model with which weights are to be shared) must be provided
-        Args
-        ----
-        layers_dim : list
-            Dimensionality of each layer
-        activation : Activation
-            Activation function
-        """
-        if self.shared_weights == True and self.base is not None:
-            mlp = nn.Sequential(
-                self.base.model[:-1],
-                nn.Linear(
-                    self.base.model[-1].in_features,
-                    self.base.model[-1].out_features,
-                    dtype=self.float,
-                ),
-            )
-            return mlp
-        elif self.shared_weights == False:
-            layers_dim = (
-                [self.state_dim] + [self.n_hid] * self.n_layers + [self.output_dim]
-            )
-            mlp = nn.Sequential(
-                *(
-                    sum(
-                        [
-                            [
-                                nn.Linear(
-                                    idim,
-                                    odim,
-                                    dtype=self.float,
-                                )
-                            ]
-                            + ([activation] if n < len(layers_dim) - 2 else [])
-                            for n, (idim, odim) in enumerate(
-                                zip(layers_dim, layers_dim[1:])
-                            )
-                        ],
-                        [],
-                    )
-                    + self.tail
-                )
-            )
-            return mlp
-        else:
-            raise ValueError(
-                "Base Model must be provided when shared_weights is set to True"
-            )
-
 
 class Policy(ModelBase):
     def __init__(self, config, env, device, float_precision, base=None):
@@ -116,6 +63,21 @@ class Policy(ModelBase):
         elif self.type == "mlp":
             self.model = self.make_mlp(nn.LeakyReLU()).to(self.device)
             self.is_model = True
+        else:
+            raise "Policy model type not defined"
+
+    def instantiate(self):
+        if self.type == "fixed":
+            self.model = self.fixed_distribution
+            self.is_model = False
+        elif self.type == "uniform":
+            self.model = self.uniform_distribution
+            self.is_model = False
+        elif self.type == "mlp":
+            from policy.mlp import MLPPolicy
+            mlp_policy = MLPPolicy(self.config, self.env, self.device, self.float_precision, self.base)
+            self.model = mlp_policy.model
+            self.is_model = mlp_policy.is_model
         else:
             raise "Policy model type not defined"
